+++ pkg/storage/tsdb/index_cache.go
@@ -13,6 +13,7 @@ import (
 	"github.com/thanos-io/thanos/pkg/model"
 	storecache "github.com/thanos-io/thanos/pkg/store/cache"
 
+	chunkCache "github.com/cortexproject/cortex/pkg/chunk/cache"
 	"github.com/cortexproject/cortex/pkg/util"
 )
 
@@ -23,6 +24,9 @@ const (
 	// IndexCacheBackendMemcached is the value for the memcached index cache backend.
 	IndexCacheBackendMemcached = "memcached"
 
+	// IndexCacheBackendMemcached is the value for the redis index cache backend.
+	IndexCacheBackendRedis = "redis"
+
 	// IndexCacheBackendDefault is the value for the default index cache backend.
 	IndexCacheBackendDefault = IndexCacheBackendInMemory
 
@@ -30,7 +34,7 @@ const (
 )
 
 var (
-	supportedIndexCacheBackends = []string{IndexCacheBackendInMemory, IndexCacheBackendMemcached}
+	supportedIndexCacheBackends = []string{IndexCacheBackendInMemory, IndexCacheBackendMemcached, IndexCacheBackendRedis}
 
 	errUnsupportedIndexCacheBackend = errors.New("unsupported index cache backend")
 	errNoIndexCacheAddresses        = errors.New("no index cache backend addresses")
@@ -40,6 +44,7 @@ type IndexCacheConfig struct {
 	Backend             string                   `yaml:"backend"`
 	InMemory            InMemoryIndexCacheConfig `yaml:"inmemory"`
 	Memcached           MemcachedClientConfig    `yaml:"memcached"`
+	Redis               chunkCache.RedisConfig   `yaml:"redis"`
 	PostingsCompression bool                     `yaml:"postings_compression_enabled"`
 }
 
@@ -53,6 +58,7 @@ func (cfg *IndexCacheConfig) RegisterFlagsWithPrefix(f *flag.FlagSet, prefix str
 
 	cfg.InMemory.RegisterFlagsWithPrefix(f, prefix+"inmemory.")
 	cfg.Memcached.RegisterFlagsWithPrefix(f, prefix+"memcached.")
+	cfg.Redis.RegisterFlagsWithPrefix(prefix, "Cache config for index in blocks storage. ", f)
 }
 
 // Validate the config.
@@ -61,10 +67,13 @@ func (cfg *IndexCacheConfig) Validate() error {
 		return errUnsupportedIndexCacheBackend
 	}
 
-	if cfg.Backend == IndexCacheBackendMemcached {
+	switch cfg.Backend {
+	case CacheBackendMemcached:
 		if err := cfg.Memcached.Validate(); err != nil {
 			return err
 		}
+	case CacheBackendRedis:
+		// redis config not provide validation
 	}
 
 	return nil
@@ -85,6 +94,8 @@ func NewIndexCache(cfg IndexCacheConfig, logger log.Logger, registerer prometheu
 		return newInMemoryIndexCache(cfg.InMemory, logger, registerer)
 	case IndexCacheBackendMemcached:
 		return newMemcachedIndexCache(cfg.Memcached, logger, registerer)
+	case IndexCacheBackendRedis:
+		return newRedisIndexCache(cfg.Redis, logger, registerer)
 	default:
 		return nil, errUnsupportedIndexCacheBackend
 	}
@@ -112,4 +123,9 @@ func newMemcachedIndexCache(cfg MemcachedClientConfig, logger log.Logger, regist
 	}
 
 	return storecache.NewMemcachedIndexCache(logger, client, registerer)
+}
+
+func newRedisIndexCache(cfg chunkCache.RedisConfig, logger log.Logger, registerer prometheus.Registerer) (storecache.IndexCache, error) {
+	client := chunkCache.NewRedisClient(&cfg)
+	return NewRedisIndexCache(client, logger, registerer)
 }